move vis_timeline into base_helper. Delete other helper file
